Use 'files' over 'executable' in 'DefaultInfo'
Executable does not make sense, it is just what the 'filegroup' rules
does, but 'files' work just as well, and have a more limited scope of
the files, suiting our code generation needs.

I do not have the scaffolding code at hand, but the gist of it was that
using the 'default' output group in the code generation rule did not
provide files that a 'cc_library' could use as sources.
Using a 'filegroup' filter that requested the output group made them
available, so whatever providers the 'filegroup' produced should be done
by the code generation rule. This was verified with the Starlark Cquery
'output_groups.cquery'. As the 'filegroup' did send them as executable I did
the same without reflection. Fredrik pointed out the error and we now
use un-executable files.